added states and events!
Added states:

IDLE
SELECT
ROLL

Select doesn't do anything now, but ROLL "unfreezes" the rigid bodies so when you first start the app, the dice are frozen initially as the title text fades.

Still a WIP in progress but definitely an improvement now that there are some states and events.